requirements.txt
absl-py==1.0.0
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.1
azure-devops==6.0.0b4
backcall==0.2.0
beautifulsoup4==4.11.1
bidict==0.22.0
boto3==1.26.105
botocore==1.29.105
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==4.4.2
exceptiongroup==1.0.1
executing==1.2.0
filelock==3.10.0
flatbuffers==1.12
fonttools==4.38.0
frozenlist==1.3.3
gast==0.4.0
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.46.3
h11==0.14.0
h5py==3.7.0
httplib2==0.22.0
huggingface-hub==0.13.3
idna==3.3
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib-metadata==4.11.4
instagrapi==1.16.29
ipykernel==6.21.1
ipython==8.9.0
isodate==0.6.1
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==8.0.2
jupyter_core==5.2.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
moviepy==1.0.3
mpmath==1.3.0
msrest==0.6.21
multidict==6.0.4
nest-asyncio==1.5.6
netifaces==0.10.6
networkx==3.0
numpy==1.22.4
oauthlib==3.2.0
openai==0.27.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.0.0
proglog==0.1.10
prompt-toolkit==3.0.36
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycryptodomex==3.15.0
pydantic==1.9.2
Pygments==2.14.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
selenium==4.6.0
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
socketio==0.2.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
statsmodels==0.13.5
sympy==1.11.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
torch==2.0.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.27.2
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.2.0
undetected-chromedriver==3.1.6
uritemplate==4.1.1
urllib3==1.26.9
vsts==0.1.25
wcwidth==0.2.6
webdriver-manager==3.8.4
websockets==10.4
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
zipp==3.8.0
